XeviGetVisualInfo: Free & clear *evi_return, not evi_return pointer
evi_return is passed in as a pointer to a location into which
XeviGetVisualInfo is expected to write a pointer to the memory
it allocated for the returned structures.  If we're failing and
bailing out, we need to dispose of the pointer we set, not the
one passed into us (which the caller may have put on the stack
or allocated as part of a larger structure).

Flagged by cppcheck 1.64:
[lib/libXext/src/XEVI.c:182] -> [lib/libXext/src/XEVI.c:186]:
 (warning) Possible null pointer dereference: evi_return - otherwise it
           is redundant to check it against null.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>